requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.0
PyJWT>=2.8,<3.0
cryptography>=41.0
requests>=2.31,<3.0
azure-identity>=1.15,<2.0
azure-keyvault-secrets>=4.7,<5.0
mysqlclient>=2.2,<3.0
python-dotenv>=1.0
gunicorn>=21.2
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3
